Write 10X hdf5 file

Write count matrix to 10X CellRanger hdf5 file. This file will be used for ArchR::import10xFeatureMatrix function.

Write10X_h5(
  seu,
  filename,
  layer = "counts",
  assay = "RNA",
  barcodes = "DNAbarcode"
)

Arguments

seu

The Seurat object after barcode remapping by replace_RNA_barcode.

filename

Path to h5 file

layer, assay

The parameter for GetAssayData.

barcodes

The colname of barcode in metadata. If it is a Function, such as colnames, The function will be used to extract the barcode. Otherwise, the column in metadata will be used as barcode.

Value

nothing

Examples

# example code